libretro-mame2014: Import libretro-mame2014-20170316 as wip/libretro-mame2014
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

MAME is the  Multiple Arcade Machine Emulator. This is based on version 0.159
with backported patches for performance and compatibility improvements.