libretro-4do: Import libretro-4do-20170423 as wip/libretro-4do
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

4DO emulates the 3DO Interactive Multiplayer games console.