py-hangups: Import py-hangups-0.4.2 as wip/py-hangups
hangups is the first third-party instant messaging client for Google
Hangouts. It includes both a Python library and a reference client
with a text-based user interface.

Unlike its predecessor Google Talk, Hangouts uses a proprietary,
non-interoperable protocol. hangups is implemented by reverse-engineering
this protocol, which allows it to support features like group
messaging that aren't available in clients that connect via XMPP.

hangups is still in an early stage of development. The reference
client is usable for basic chatting, but the API is undocumented
and subject to change.