libretro-bsnes-mercury: Import libretro-bsnes-mercury
Libretro is a simple but powerful development interface that allows
for the easy creation of emulators, games and multimedia applications
that can plug straight into any libretro-compatible frontend. This
development interface is open to others so that they can run these
pluggable emulator and game cores also in their own programs or
devices.

bsnes-mercury is a Super Nintendo Entertainment System / Super
Famicom video game system emulator that can be used as a libretro
core.

It is a fork of bsnes aiming to restore some useful features that
have been removed, as well as improving performance a bit. Maximum
accuracy is still uncompromisable; anything that affects accuracy
is optional and off by default.

It has has 3 different profiles: Performance, Balanced, and Accuracy.
The Balanced version is recommended for modern PCs. Use the Accuracy
profile only for extreme cases. For more information, see:

https://wiki.libretro.com/index.php?title=Nintendo_SNES_Core_Compatibility