libretro-mednafen-vb: Import libretro-mednafen-vb-20170501 as wip/libretro-mednafen-vb
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

Beetle/Mednafen VB is a Nintendo Virtual Boy video game system emulator that
can be used as a libretro core.