libretro-tyrquake: Import libretro-tyrquake-20170410 as wip/libretro-tyrquake
Libretro is a simple but powerful development interface that allows
for the easy creation of emulators, games and multimedia applications
that can plug straight into any libretro-compatible frontend. This
development interface is open to others so that they can run these
pluggable emulator and game cores also in their own programs or
devices.

Quake is an action game by id Software released in 1996. It's a
shooter similar in style to Doom, but featuring a fully 3d environments
and enemies. The ability to play networked games with other players
immediately made the game very popular with the online community.
Later, id Software released a freely downloadable update specifically
targeted at improving play on the internet, even over slow modem
connections. This was known as "QuakeWorld" and hundreds of servers
popped up all over the place.

TyrQuake is an interpreter for Quake and Quakeworld that can be
used as a libretro core.