libretro-prboom: Import libretro-prboom-20170410 as wip/libretro-prboom
Libretro is a simple but powerful development interface that allows
for the easy creation of emulators, games and multimedia applications
that can plug straight into any libretro-compatible frontend. This
development interface is open to others so that they can run these
pluggable emulator and game cores also in their own programs or
devices.

PrBoom is a Doom source port that can be used as a libretro core.