libretro-fmsx: Import libretro-fmsx-20170223 as wip/libretro-fmsx
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

fMSX is a program that emulates MSX, MSX2, and MSX2+ 8bit home computers.