surf2: Import surf2-2.0 as wip/surf2
surf is a simple web browser based on WebKit2/GTK+. It is able to
display websites and follow links. It supports the XEmbed protocol
which makes it possible to embed it in another application.
Furthermore, one can point surf to another URI by setting its
XProperties.

XXX: This is imported as surf2 and program and man page renamed accordingly
XXX: in order to avoid conflicts with wip/surf. When wip/webkit-gtk and this
XXX: package will be imported in pkgsrc probably it's better to rename it to
XXX: just surf given that upstream now maintains only the WebKit2/GTK+ version
XXX: and it seems that wip/webkit-gtk is in better shape than previous versions.