ioquake3-snapshot: import ioquake3-20170521 as wip/ioquake3-snapshot
ioquake3 is a game engine project which aims to build upon id Software's Quake
3 source code release. The source code was released on August 20, 2005 under
the GPL.

The project's goal is to create a bug-free, enhanced open source Quake 3
source code distribution upon which new games and projects can be based, as
well as an improved version (in relation to the original binaries) of the
engine for playing Quake III: Arena, the Team Arena expansion pack and all the
popular mods.

This package is a snapshot of the development version of games/ioquake3.

See also:
https://ioquake3.org/2017/03/13/important-security-update-please-update-ioquake3-immediately/