Add cpluff
C-Pluff is a plug-in framework for C programs.

It has been strongly inspired by the Java plug-in framework in
Eclipse. C-Pluff focuses on providing core services for plug-in
interaction and plug-in management. It aims to be platform neutral
and supports dynamic changes to plug-in configuration without
stopping the whole application or framework.