libretro-pcsx-rearmed: Import libretro-pcsx-rearmed-0.22 from pkgsrc as wip/libretro-pcsx-rearmed
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

PCSX ReARMed is a Sony PlayStation1 video game system emulator.