Import py27-llvm-lit-5.0.0nb20170702 as wip/py-llvm-lit-git.
lit is a portable tool for executing LLVM and Clang style test suites,
summarizing their results, and providing indication of failures. lit is
designed to be a lightweight testing tool with as simple a user interface as
possible.

lit should be run with one or more tests to run specified on the command line.
Tests can be either individual test files or directories to search for tests.