gargoyle: add package based on Wes Frazier's version from PR 52431
Gargoyle is an IF player that supports all the major interactive fiction
formats.

Most interactive fiction is distributed as portable game files. These portable
game files come in many formats. In the past, you used to have to download a
separate player (interpreter) for each format of IF you wanted to play.

Gargoyle is based on the standard interpreters for the formats it supports.
Gargoyle is free software released under the terms of the GNU General Public
License.

TODO: does not link on NetBSD.