nmh: Import nmh-1.7rc1 as wip/nmh
nmh (new MH) is a powerful electronic mail handling system. It was
originally based on version 6.8.3 of the MH message system developed by
the RAND Corporation and the University of California. It is intended
to be a (mostly) compatible drop-in replacement for MH.

nmh consists of a collection of fairly simple single-purpose programs
to send, receive, save, retrieve, and manipulate e-mail messages. Since
nmh is a suite rather than a single monolithic program, you may freely
intersperse nmh commands with other commands at your shell prompt,
or write custom scripts which use these commands in flexible ways.

Based on mail/nmh with few adjustments that will be needed for the next nmh-1.7
release.

Please note this is a RC release and it is intended to ease testing and updating
of mail/nmh!