refkeen: update to refkeen-20170814
Changelog
---------

Aug 14, 2017 (v0.18.0):
* Move some memory management code from the differing id_mm.c files to a
separate be_cross_mem.c file. For the sake of simplicity, XMS and EMS are
gone. This newly added source file also grants us some more control over
memory management, using functions like BE_Cross_Bfarmalloc as replacements
for functions originally used in the DOS sources (such as farmalloc).
* Somewhat better support for multi-EXE emulation, like CATABYSS.EXE
and INTRO.EXE for The Catacomb Abyss v1.13. This includes a replacement
for the execv function, reverting the state of the runtime stack.
This is mostly useful when two DOS EXEs may execute each other, over and
over again (e.g., INTRO.EXE and DEMOCAT.EXE from The Catacomb Abyss v1.13).
* Minor loadscn changes (mostly renames).
* An almost complete port of DEMOCAT.EXE/HINTCAT.EXE is now in. It's still
impossible to send anything to a printer, but otherwise it should have
similar behaviors (albeit the sound beep periods may be shorter).
As a side-note, this makes the 'F10' key usable in
the intro, with The Catacomb Abyss v1.13.
* Minor related fixes in be_st_sdl_audio_timer.c and be_st_sdl_graphics.c,
as well as small changes for alt. controller mappings handling.
* Also fix parsing of filenames from e.g., SCRIPT.HNT, The Catacomb Armageddon
(for instance, "HINT01.HNT " should first be trimmed to "HINT01.HNT").
* For the later Catacombs titles, it's now possible to choose between
the game and HINTCAT.EXE/DEMOCAT.EXE, right from the launcher. The last
choice is saved for the next time Reflection Keen is started. There's
further the addition of the related /slidecat command-line argument.
* More changes for game versions management - We now store a separate list
of "main functions" for each game version. Usually, they have matching
EXE filenames (Keen Dreams 2015 is an exception).
* When Reflection Keen checks for an available game installation, all it does
is check for recognized game data (and EXE) files. It does not, at this point,
look for embedded data, like INTROSCN.SCN in CAT3D.EXE. It also does not
unpack any packed EXE. This may only happen when the game (or slidecat)
is started, *or* when BE_Cross_Bexecv is called.
* In particular, embedded data like INTROSCN.SCN is *not* written as
an external file anymore. It's also impossible to use replacements for
these with "manualgamevermode=true".
* Ensure last display number. is saved, regardless of any specific event.
* On an attempt to resolve the issue of high audio latency leading to a low
framerate (reproduced on Android), we now basically run the various sound
callbacks (originally SDL callbacks) in the main thread. This is done in
virtually the same way as if the sound system were disabled, only that
an additional buffer is used for passing samples to the SDL audio thread.
* The new "sndinterthreadbufferratio" setting can be used to adjust this
buffer's size, as well as the sizes of internal PC and AL buffers.
They're calculated relatively to the buffer size chosen for
the SDL callback thread (used only in this callback).
* Other misc. fixes.