dasht: Import dasht-2.2.0 as wip/dasht
dasht is a collection of shell scripts for searching, browsing,
and managing API documentation (in the form of 150+ offline
documentation sets, courtesy of Dash for OS X) all from the comfort
of your own terminal!

The name "dasht" is a portmanteau of Dash and the letter "t", for
terminal.  Etymologically, "dasht" is Persian for plain, as in an
flat expanse of land, which aptly characterizes the terminal
environment where everything is text.