TODO: delete an entry about moreutils-0.61
moreutils is now in pkgsrc and maybe we will need to add an entry to
doc/TODO about that. It wasn't updated to version 0.61 because the distfile
used are from Debian (this is also suggested by upstream, that doesn't
officially provide any tarballs). At least for 0.61 according the changelog
there isn't any significant changes and so that's maybe why the Debian
maintainers haven't updated moreutils there (so, probably maintaining the
tarballs and uploading them to MASTER_SITE_LOCAL isn't worth the effort right
now).