libretro-np2kai: Import libretro-np2kai-0.11
Libretro is a simple but powerful development interface that allows for
the easy creation of emulators, games and multimedia applications that
can plug straight into any libretro-compatible frontend. This development
interface is open to others so that they can run these pluggable emulator
and game cores also in their own programs or devices.

Neko Project II is a PC-9801 emulator. The PC-9800 series, commonly
shortened to PC-98, is a lineup of Japanese 16-bit and 32-bit personal
computers manufactured by NEC from 1982 through 2000.

Despite hardware specifications far inferior to the Fujitsu FM Towns
and Sharp X68000 personal computers, the massive install base and steady
flow of game titles (in particular "dojin" style dating sims and RPGs,
as well as early games of the Touhou Project franchise) kept the PC-9801
as the favored platform for PC game developers in Japan until the rise
of the DOS/V clones.