Import blosc-1.2.1
Blosc is a high performance compressor optimized for binary data. It
has been designed to transmit data to the processor cache faster than
the traditional, non-compressed, direct memory fetch approach via a
memcpy() OS call. Blosc is the first compressor that is meant not only
to reduce the size of large datasets on-disk or in-memory, but also to
accelerate memory-bound computations.