Import fonttools version 3.15.1
fontTools is a library for manipulating fonts, written in Python. The project
includes the TTX tool, that can convert TrueType and OpenType fonts to and from
an XML text format, which is also called TTX. It supports TrueType, OpenType,
AFM and to an extent Type 1 and some Mac-specific formats. The project has a
BSD-style open-source licence.

Remaining issues before importing to pkgsrc:
- the path to the manual page;
- the license.