compiler-rt-netbsd: Disable background thread on NetBSD
This thread does g/c of shadow memory and writes heap profile
periodically. It's not required for operation for real programs now.

It breaks because it attempts to create new POSIX thread before
initialization of libc and libpthread startup code.

Once tsan will pass all tests on NetBSD, this can be implemented differently
to support the NetBSD case defering startup of this thread later.

Sponsored by <The NetBSD Foundation>