tig: Import tig-2.3.0 as wip/tig
Tig is a git repository browser that additionally can act as a pager for
output from various git commands.

When browsing repositories, it uses the underlying git commands to present
the user with various views, such as summarized revision log and showing
the commit with the log message, diffstat, and the diff.

Using it as a pager, it will display input from stdin and colorize it.

This package is intended as an update for devel/tig, it was only briefly tested
and some (new) patches needs a review. Please look at the TODO for more
information about that, any possible review/feedback/critics/testing (especially
from devel/tig users) are welcome!

About pkgsrc changes related to devel/tig:
 - patch-tig_c is no more needed, the logic in tig-2.3.0 is different
 - patch-contrib_config.make-Darwin: already applied by upstream, no more needed