edbrowse: Update wip/edbrowse to 3.7.1
pkgsrc changes:
 - Use MASTER_SITE_GITHUB for MASTER_SITES and just fetch the .tar.gz archives
   there
 - Depends on lang/libduktape and drop dependency to wip/spidermonkey24

Changes:
3.7.1:
Automatically expand a frame when javascript tries to access its objects.
Otherwise it won't expand unti you expand it explicitly.
This saves resources; don't fetch and expand a frame unless it is needed.

^c interrupts javascript, if that javascript calls our native
methods. It also interrupts http fetch.  Hit ^c if the browsing is
taking too darn long - however - the page may not render properly
or completely.  Don't do this if you are using this page for
ecommerce or something vital.

Keep edbrowse in 1 process. Don't fork a javascript process to make
all the js calls.  There are many technical reasons why 1 process
is a better design.

The keyword jspool is no longer used in the .ebrc config file.  This
was needed for Mozilla JS, but it is not needed for Duktape.

Gather keyword value pairs from <style> tags and css files and
apply them to the corresponding javascript objects.  Implement
getComputedStyle().

3.7.0:
Switch from Mozilla js to Duktape js.

Mask password fields on input forms with stars, as other browsers do.
Issue the ipass command to enter a password without echo.

Curl authorization negotiations enabled or disabled via the can toggle command.
This is a workaround for problems with NTLM.

Parallel instances of edbrowse don't clobber each other's cookies
when they exit and write the common cookie jar.  See mergeCookies()
in cookies.c.  curl does not become active until you need it.

Environment variable JSGC forces duktape garbage collection after
every script.  Environment variable JS1 keeps edbrowse and js in
one process.  These are for development and testing, and could go
away.

Clean up compiler warnings so we can use the -Wall flag.

3.6.3:
Maintain a cache of http files.
Provides faster access to websites that are visited often.
Cache directory and size can be set in the config file, but the
defaults are reasonable.

Use a substring of the url to determine a mime type, not just the suffix.
This is primarily for youtube videos, which can play as a stream, but have no
obvious protocol or suffix to key on.
urlmatch = youtube.com/watch?|youtube.com/embed/
Currently a substring, may become a regexp later.

g? prints the url for the link instead of going to it. You can look before you
leap.
Also g2? g$? etc.

i* on a textarea <buffer 7> goes to that editing session, in this
case equivalent to e7.  This is convenient for scripting, since
you don't know ahead of time which buffer will be allocated for
this purpose.

The M (move) command does not require a destination session; edbrowse
finds an empty session for you.

Represent multiple frames in one edbrowse window. Each has its own javascript
world.  At this point the worlds do not interact.
You can expand and contract frames by the exp and ctr commands.