freedoom: Update wip/freedoom to 0.11.3
Changes:
2017-07-18: Freedoom 0.11.3 released

This point release is only to allow building from source with DeuTex 5.0, and
thus is considered unnecessary for any purpose other than distribution
packaging. There are no level, sprite, sound, music, or other asset changes
from 0.11.2.

2017-03-15: Freedoom 0.11.2 released

Solving a few more issues found in the previous release, this point release is
more dry than the last one, but we hope that it just refines the game to make
it better than before.

  * Missing multiplayer starts have been added to several levels. No map should
    be lacking anymore.
  * A few mapping errors were repaired by changing sector heights. Speedrunners
    can better appreciate smooth flow through the levels.
  * Easy and normal difficulty levels have been tweaked.
  * Par times for Phase 1 have been added.
  * Cleaned up title screen using the Freedoom font for Phase 1 and Phase 2 on
    screen.

2017-02-22: Freedoom 0.11.1 released

We've had quite a good reception for 0.11 and are thankful to everyone sending
feedback, both praise and criticism, we use it all for improving Freedoom
further! This point release repairs a few deficiencies found in the 0.11
release, and adds new resources.

  * 1% armor bonus pickups are recolored from red to green. They were too
    easily confused for health pickups.
  * New TNT: Evilution-compatible textures.
  * New pain bringer and pain lord sprites, replacing old concept art-derived
    ones.
  * A few map errors are fixed, including misaligned textures, leftover Boom
    specials, and it should no longer be possible to get stuck between a rock
    and a tree in Phase 1 C1M1.
  * Brand new Phase 1 C3M1, replacing an old Doom-inspired level.
  * Widescreen statusbar for ZDoom was removed. This created incompatibility
    with some mods, and mod compatibility is prefered over visual enhancements.
  * Fixed intermission text after using the secret exit in Phase 2 MAP31, where
    obituary texts were incorrectly displayed instead.
  * More new music tracks, providing better vanilla compatibility.

2017-02-16: Freedoom 0.11 released

Another year, and one of Freedoom's biggest updates has arrived! One of the
major developments we've had is the elimination of Boom as the target engine
for game support. While we are aiming for an eventual compatibility with
Vanilla Doom, Freedoom should presently run on any limit-removing engine, which
is nearly all of them.

  * Lots of new music.
  * New levels, including C1M1 kicking off Phase 1.
  * Aquatex: over 200 new textures for mappers to use.
  * New intermission screens.
  * Some new weapon sprites: new pistol, new pickups.
  * New medkit and armor pickup sprites.
  * New project logo.