omnispeak: import omnispeak-1.0 as wip/omnispeak
DESCR:

Omnispeak is an open-source reimplementation of Commander Keen
episodes 4, 5, and 6. It aims to be a pixel-perfect, bug-for-bug
clone of the original games, and is compatible with savegames from
the DOS version.

The game data files are required to run the game (shareware or
registered copy).

PKGSRC changes: Search for common data files in the package's
data directory if not found in the current directory.