Remove xombrero.
Thomas Mueller on pkgsrc-users:

Quoting from Wikipedia:

In early 2017, it appeared the project was likely to be discontinued, as
xombrero depended on an old version of WebKit which had multiple security
vulnerabilities, and a port to a modern version of WebKit would be difficult.
Due to these issues, OpenBSD removed xombrero from its ports tree on 1 February
2017.

Github page says

xombrero has been retired and is no longer under development and supported.