libretro-melonds: Import libretro-melonds-20171207
Seems to play Tetris DS ok. It is also PaX-friendly, unlike libretro-desmume.

Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

melonDS is a Nintendo DS emulator with a goal of doing things "right and fast".

Note: Currently the libretro core lacks touchscreen/mouse support.