libretro-cap32: Import 20171109
Libretro is a simple but powerful development interface that allows for
the easy creation of emulators, games and multimedia applications that
can plug straight into any libretro-compatible frontend. This development
interface is open to others so that they can run these pluggable emulator
and game cores also in their own programs or devices.

Caprice32 is an Amstrad CPC emulator.

It is recommended that users enable "autorun" in the Core Options in
the RetroArch F1 menu after selecting a disk image for the first time,
then restart the core to load it.

A virtual keyboard can be toggled by pressing the virtual "Y" button.