libretro-beetle-saturn: Update to 20171209
hiddenasbestos brought the libretro port up to parity with upstream
v0.9.48, which gives a few ton of nice accuracy improvements and the
much-awaited savestate support (savestates made with the v0.9.47
version are not compatible with the v0.9.48 update, so beware if
you made a bunch of states in the short time between updates).
Tatsuya79 and retro-wertz provided a number of improvements to the
overscan and cropping behavior.

Button input was previously all-digital, now button input can be
analog as well. As a proof of concept, this has already been
implemented for the Beetle Saturn core. Analog triggers now work
as expected. This feature will be necessary for future systems like
PlayStation2, where each face button on the gamepad was an analogue
button.

Mouse buttons 4 and 5 were added. A proof of concept has already been
implemented for Beetle Saturn. The mouse on the Sega Saturn had at
least 5 buttons instead of the PlayStation mouse’s 2.

Support for multiple disc selection, lightguns, and an emulated
mouse sensitivity option.

Probably more stuff that I've missed.