zathura-pdf-mupdf: Import zathura-pdf-mupdf--0.3.1nb7 as wip/zathura-pdf-mupdf
The zathura-pdf-mupdf plugin adds PDF support to zathura by using the
mupdf rendering library.

This package is intended as a WIP update to print/zathura-pdf-mupdf due API
changes in mupdf-1.12.0.

TODO: Still doesn't work properly and crashes on exit, patches to adjust to the
TODO: new API aren't commented and still need to be reviewed.