farbfeld: Update wip/farbfeld to farbfeld-3
pkgsrc changes:
 - Clean up a bit the pkgsrc Makefile and buildlink3.mk
 - Use MAKE_FLAGS to inject variables and delete patches/patch-config.mk

Changes (from email announcements):
farbfeld-3:
The main focus of this release was stability and code refactorizations.
Fortunately, it all worked out very well and across the board you can
count on these tools not to fail on you.
The build system was revamped, heavily simplified and is with one small
exception POSIX compliant. I also reworked the invert.c example filter
to reflect the larger focus on hardening the tools.
I fixed quite a few subtle edge-cases in the tools and the scripts, and
now they should all work flawlessly. If not, please feel free to report
bugs!

farbfeld-2:
 - 75+% speed improvement in the conversion tools
   by handling row buffers instead of each channel
   individually with atomic writes.
 - Use convert instead of xconvert in 2ff(1) to
   actually not error out when we hit anything other
   than png and jpg.
 - Major code cleanup and documentation improvement.