nmh: Import nmh-1.7.1rc2 as wip/nmh
nmh (new MH) is a powerful electronic mail handling system. It was
originally based on version 6.8.3 of the MH message system developed by
the RAND Corporation and the University of California. It is intended
to be a (mostly) compatible drop-in replacement for MH.

nmh consists of a collection of fairly simple single-purpose programs
to send, receive, save, retrieve, and manipulate e-mail messages. Since
nmh is a suite rather than a single monolithic program, you may freely
intersperse nmh commands with other commands at your shell prompt,
or write custom scripts which use these commands in flexible ways.

This is intended to ease the update of mail/nmh.