hub: Use lang/go/go-package.mk and inject `do-install'
Instead of doing post-extract/do-build and other Go contortions
just reuse all the functionalities provided by `.../../lang/go/go-package.mk'.
This also simplify the Makefile a lot.

If we do not define do-install a lot of not needed files are automatically
installed (and usually this is not desiderable for Go packages that just provide
a binary). Inject do-install accordingly and also manually install the man page.

Thanks a lot to <bsiegert> for kindly pointing out that and for
suggesting to see www/caddy as example!