mitmproxy: Import mitmproxy-3.0.0 as wip/mitmproxy
This package contains tool for HTTP(S) sniffing.

mitmproxy is an interactive, SSL-capable man-in-the-middle proxy
for HTTP with a console interface.

mitmdump is the command-line version of mitmproxy. Think tcpdump
for HTTP.

libmproxy is the library that mitmproxy and mitmdump are built on.

XXX: Completely based on net/mitmproxy and intended to update it. Apart
XXX: some minor review and a bit more testing it is probably ready to
XXX: be used to update mitmproxy package.