libretro-freeintv: Import libretro-freeintv-20170215
Libretro is a simple but powerful development interface that allows for
the easy creation of emulators, games and multimedia applications that
can plug straight into any libretro-compatible frontend. This development
interface is open to others so that they can run these pluggable emulator
and game cores also in their own programs or devices.

FreeIntv is a libretro emulation core for the Mattel Intellivision
designed to be compatible with joypads from the SNES era forward even
if they originally required a number pad.