mk/git-package.mk: optimize for single-repository case
In the majority of cases, a package corresponds to a single Git
repository. Therefore, having three separate lines in the package
Makefile is only overhead. The only line that such a package needs is

    GIT_REPO=       git://git@github.com/NetBSD/pkgsrc

This automatically sets GIT_REPOSITORIES, GIT_MODULE and WRKSRC.

This change is completely backwards-compatible, allowing the affected
packages to be migrated on a case-to-case basis.