mk/*-package.mk: provide show-all targets for the VCS include files
For understanding the mk files for the version control systems, it is
helpful to get a concise list of the input and output variables. This is
done by the show-all target, or more specialized by show-all-cvs,
show-all-git, etc.

Having this information condensed in a single place also helps
implementing new features, like optimizing for packages that only need a
single repository. In many cases this can be configured with a single
variable, instead of currently 4.