mk/*-package.mk: test packages needing multiple repositories
A package can fetch from multiple repositories at the same time. It is
even possible to mix different SCM systems.

While testing this, it became clear that the word MODULE in the variable
names only applies to CVS. In all other SCMs, it was a copy-and-paste
mistake. It has therefore been changed to EXTRACTDIR, since that
describes the intention much better. For Git, the old GIT_MODULE will
continue to work for a short time, the other SCMs don't need this
backwards compatibility since only a single package used that feature.

The file name for the cached archive has been changed to include the
repository ID. This was necessary to support multiple checkouts of the
same repository.