mk/*-package.mk: replace ad-hoc tests with Kyua tests
Before, the tests were all written in a single file, which was not
extensible and difficult to understand.

When run directly, the tests all pass. When run via Kyua, they fail
because they expect the file test.subr in the same directory. Someone
familiar with Kyua needs to fix this.