ez-pine-gpg: add new package
ez-pine-gpg is a set of scripts that allows beginners and experts
to use gpg with Pine. There are plenty of other applications that
allow gpg to be used with Pine: This one is intended to be the
best, since it merges intuitive use with powerful features. The
result is an application that's not only fast and secure, but also
perfect for novices and power-users alike.