vimb3: import vimb3-3.1.0 as wip/vimb3
This is the latest release of wip/vimb3-git.

vimb is a fast and lightweight web browser based on the webkit web browser
engine and the gtk toolkit. vimb is modal like the great vim editor and
also easily configurable during runtime. vimb is mostly keyboard driven
and does not detract you from your daily work.