gst-plugins1-hls: import gst-plugins1-hls-1.14.0 as wip/gst-plugins1-hls
GStreamer is a library that allows the construction of graphs of
media-handling components, ranging from simple Ogg/Vorbis playback to
complex audio (mixing) and video (non-linear editing) processing.

Applications can take advantage of advances in codec and filter technology
transparently.  Developers can add new codecs and filters by writing a
simple plugin with a clean, generic interface.

GStreamer is released under the LGPL.

This package is part of the 'bad' plugins for GStreamer.  It provides a
plugin for "hls".